добавляет динамическую форму входа
